separated doc utils from other utils
