+++ src/NCase.sln
@@ -55,6 +55,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NCase.XunitAdapter", "NCase
 		{BADB59FB-1A14-42AF-B314-152C4787F437} = {BADB59FB-1A14-42AF-B314-152C4787F437}
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NDocUtil", "NDocUtil\NDocUtil.csproj", "{E95016D2-61FD-4884-AC80-3630F54F30CC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -105,6 +107,10 @@ Global
 		{A04648E1-25C7-41FB-A2B9-61D5D407566B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A04648E1-25C7-41FB-A2B9-61D5D407566B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A04648E1-25C7-41FB-A2B9-61D5D407566B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E95016D2-61FD-4884-AC80-3630F54F30CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E95016D2-61FD-4884-AC80-3630F54F30CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E95016D2-61FD-4884-AC80-3630F54F30CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E95016D2-61FD-4884-AC80-3630F54F30CC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
