+++ src/NDocUtil/ConsoleRecord.cs
@@ -1,7 +1,7 @@
 using System;
 using JetBrains.Annotations;
 
-namespace NUtil.Doc
+namespace NDocUtil
 {
     public class ConsoleRecord
     {
