+++ src/NDocUtil/Properties/AssemblyInfo.cs
@@ -1,0 +1,17 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following 
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+
+[assembly: AssemblyTitle("NDocUtil")]
+[assembly: AssemblyProduct("NCase")]
+[assembly: AssemblyDescription("Utilities to write documentation on C# Api")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+
+[assembly: ComVisible(false)]
+[assembly: Guid("2fff8a7d-beb6-4c2f-93da-4ca16e5b6598")]
