+++ src/NDocUtil/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="JetBrains.Annotations" version="9.2.0" targetFramework="net45" />
+</packages>
